Paso metodos de ver y guardar imagen a utilities
