+++ src/Components/DishForm.js
@@ -2,7 +2,6 @@ import React from "react";
 import Axios from "axios";
 import { useState, useEffect } from "react";
 import { Form, Row, Col, Button, Spinner } from "react-bootstrap";
-import { saveAs } from "file-saver";
 import IngredientsDropdown from "./IngredientsDropdown";
 import Utilities from "./Common/Utilities";
 import 'bootstrap/dist/css/bootstrap.min.css';
@@ -90,18 +89,6 @@ const DishForm = ({ initialValues, onSubmit }) => {
         setFormData({ ...formData, headerImage: name, base64Header: base64 });
     };
 
-    const openImage = (src) => {
-        let image = new Image();
-        image.src = src;
-
-        let w = window.open("");
-        w.document.write(image.outerHTML);
-    };
-
-    const downloadImage = (src, name) => {
-        let file = Utilities.convertBase64ToFile(src, name);
-        saveAs(file, name);
-    }
     const renderIngredients = formData.ingredients.map(ingredient => <IngredientsDropdown
         key={ingredient.listId}
         categories={categories}
@@ -243,7 +230,7 @@ const DishForm = ({ initialValues, onSubmit }) => {
                                 <div className="d-grid gap-2">
                                     <Button
                                         variant="success"
-                                        onClick={() => openImage(formData.base64Image)}
+                                        onClick={() => Utilities.openImage(formData.base64Image)}
                                     >See Image</Button>
                                 </div>
                             </Col>
@@ -251,7 +238,7 @@ const DishForm = ({ initialValues, onSubmit }) => {
                                 <div className="d-grid gap-2">
                                     <Button
                                         variant="success"
-                                        onClick={() => downloadImage(formData.base64Header, formData.headerImage)}
+                                        onClick={() => Utilities.downloadImage(formData.base64Header, formData.headerImage)}
                                     >Download</Button>
                                 </div>
                             </Col>
@@ -275,7 +262,7 @@ const DishForm = ({ initialValues, onSubmit }) => {
                                 <div className="d-grid gap-2">
                                     <Button
                                         variant="success"
-                                        onClick={() => openImage(formData.base64Header)}
+                                        onClick={() => Utilities.openImage(formData.base64Header)}
                                     >See Image</Button>
                                 </div>
                             </Col>
@@ -283,7 +270,7 @@ const DishForm = ({ initialValues, onSubmit }) => {
                                 <div className="d-grid gap-2">
                                     <Button
                                         variant="success"
-                                        onClick={() => downloadImage(formData.base64Header, formData.headerImage)}
+                                        onClick={() => Utilities.downloadImage(formData.base64Header, formData.headerImage)}
                                     >Download</Button>
                                 </div>
                             </Col>
